Initial commit
The resume_template is a reference that served as inspiration, but the
real stuff is all in blank_resume. This is my custom resume, but
really it needs to be split into a separate class from which a
document can then be created. This will make the code for generating
the document cleaner.