Changes empty circles to be unfilled
Modifies the symbol used for an empty level. By leaving it a hollow
circle, it does a better job of representing the fact that it's meant
to be filled in, and less weight draws the eyes to the measured
portion. Also modifes it so that it's the same as the accent color.

This approach has the benefit of working regardless of color, so a
black and white copy of the resume will always show the true meaning
encoded.